Deployment/requirements.txt
streamlit
seaborn
matplotlib.pyplot
numpy
joblib
scikit-learn == 1.0.2